requirements.txt
flask_cors
SQLAlchemy==2.0.20
peewee==3.16.2
python-dotenv==0.21.1
pymysql
Flask
mailjet_rest
flask_jwt_extended
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
jsonschema
PyJWT==2.6.0
peewee_enum_field
psycopg2-binary